Refactor graph.js into modular structure
- Decomposed graph.js (761 lines) into 6 focused modules
  - initialization: Handles graph setup and basic configuration
  - forceManagement: Controls force-directed graph physics
  - dataLoading: Manages API data fetching and processing
  - interactions: Handles hover, click, and context menu events
  - dragHandling: Manages node dragging and link creation
  - events: Handles event bus communication

Each module has a clear single responsibility, improving maintainability
while preserving the original API through re-exports.
